[Feat] #24 - Move Enumeration (UserRole)
